REFACTOR 찢기는 부분 계산, 별도 함수로 분리
